Revised frame counting to explicitly use render loop for updates. Revised method used to wait for next mumble tick.

+++ Blish HUD/GameServices/Debug/FrameCounter.cs
@@ -1,18 +1,18 @@
 ﻿namespace Blish_HUD.Debug {
     public class FrameCounter {
 
-        public float CurrentAverage { get; private set; }
+        public double CurrentAverage { get; private set; }
 
-        private readonly RingBuffer<float> _fpsSamples;
+        private readonly RingBuffer<double> _fpsSamples;
 
         public FrameCounter(int sampleCount) {
-            _fpsSamples = new RingBuffer<float>(sampleCount);
+            _fpsSamples = new RingBuffer<double>(sampleCount);
         }
 
-        public void Update(float deltaTime) {
-            _fpsSamples.PushValue(1 / deltaTime);
+        public void Update(double deltaTime) {
+            _fpsSamples.PushValue(1d / deltaTime);
 
-            float total = 0;
+            double total = 0;
             for (int i = 0; i < _fpsSamples.InternalBuffer.Length; i++) {
                 total += _fpsSamples.InternalBuffer[i];
             }
